Embeddable task list no toggling items complete status successfully.

+++ apps/fluster/src-tauri/src/features/task_manager/commands/create_task.rs
@@ -40,6 +40,8 @@ pub async fn create_task(task: TaskModel, tags: Vec<TaskTagModel>) -> FlusterRes
         }
     }
     TaskTagEntity::create_many(&db, tags).await?;
-    TaskTagEntity::delete_by_tag_values(&db, tag_values_to_remove).await?;
+    if !tag_values_to_remove.is_empty() {
+        TaskTagEntity::delete_by_tag_values(&db, tag_values_to_remove).await?;
+    }
     Ok(())
 }
